Place each day's image in their own folder.
Or prefix, as S3 would have it.

Way faster to `aws s3 sync` the directories now. Seems like the sync
logic is implemented client-side, so syncing the flat directory of
compressed images required a lot of bandwidth as the CLI tool pulled
down all the metadata of all of the thousands of images already in the
bucket, even if there were only a few hundred locally. Splitting the
collection into subfolders by day and syncing only those days which
exist locally is much, much faster (difference of minutes). And will
hopefully stop eating all of my throughput cap.